Move LRUMap (and ItemPool and PriorityQueue) to a new utils/dataStructures folder
