+++ src/ts/frontend/starmap/stellarPathfinder.ts
@@ -20,8 +20,8 @@ import { Vector3 } from "@babylonjs/core/Maths/math.vector";
 import { StarSystemCoordinates, starSystemCoordinatesEquals } from "@/backend/universe/starSystemCoordinates";
 import { StarSystemDatabase } from "@/backend/universe/starSystemDatabase";
 
+import { PriorityQueue } from "@/utils/dataStructures/priorityQueue";
 import { getNeighborStarSystemCoordinates } from "@/utils/getNeighborStarSystems";
-import { PriorityQueue } from "@/utils/priorityQueue";
 import { err, ok, Result } from "@/utils/types";
 
 type Node = {
