+++ src/ts/frontend/terrain/sphere/chunkForgeCompute.ts
@@ -29,6 +29,7 @@ import {
 import { HeightMap1x1, HeightMap2x4 } from "@/frontend/assets/heightMaps";
 import { IPlanetHeightMapAtlas } from "@/frontend/assets/planetHeightMapAtlas";
 
+import { LRUMap } from "@/utils/dataStructures/lruMap";
 import { Direction } from "@/utils/direction";
 import { err, ok, PowerOfTwo, Result } from "@/utils/types";
 
@@ -36,7 +37,6 @@ import { Settings } from "@/settings";
 
 import { SquareGridIndicesComputer } from "../squareGridIndexComputer";
 import { SquareGridNormalComputer } from "../squareGridNormalComputer";
-import { LRUMap } from "./lruMap";
 import { SphericalHeightFieldBuilder1x1 } from "./sphericalHeightFieldBuilder1x1";
 import { SphericalHeightFieldBuilder2x4 } from "./sphericalHeightFieldBuilder2x4";
 import { SphericalProceduralHeightFieldBuilder } from "./sphericalProceduralHeightFieldBuilder";
