+++ src/ts/frontend/universe/planets/telluricPlanet/terrain/chunks/workerPool.ts
@@ -15,7 +15,7 @@
 //  You should have received a copy of the GNU Affero General Public License
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import { PriorityQueue } from "@/utils/priorityQueue";
+import { PriorityQueue } from "@/utils/dataStructures/priorityQueue";
 
 import { BuildTask } from "./taskTypes";
