+++ src/ts/frontend/universe/stellarObjects/neutronStar/neutronStar.ts
@@ -39,8 +39,8 @@ import { defaultTargetInfoCelestialBody, TargetInfo } from "@/frontend/universe/
 import { AsteroidField } from "@/frontend/universe/asteroidFields/asteroidField";
 
 import { Cullable } from "@/utils/cullable";
+import { ItemPool } from "@/utils/dataStructures/itemPool";
 import { isSizeOnScreenEnough } from "@/utils/isObjectVisibleOnScreen";
-import { ItemPool } from "@/utils/itemPool";
 import { getRgbFromTemperature } from "@/utils/specrend";
 import { getOrbitalObjectTypeToI18nString } from "@/utils/strings/orbitalObjectTypeToDisplay";
 import { DeepReadonly } from "@/utils/types";
