+++ src/ts/playgrounds/gasPlanet.ts
@@ -29,7 +29,7 @@ import { GasPlanet } from "@/frontend/universe/planets/gasPlanet/gasPlanet";
 
 import { Settings } from "@/settings";
 
-import { ItemPool } from "../utils/itemPool";
+import { ItemPool } from "../utils/dataStructures/itemPool";
 import { enablePhysics } from "./utils";
 
 export async function createGasPlanetScene(
