+++ src/ts/playgrounds/sol/saturn.ts
@@ -28,7 +28,7 @@ import { ShadowPostProcess } from "@/frontend/postProcesses/shadowPostProcess";
 import { AsteroidField } from "@/frontend/universe/asteroidFields/asteroidField";
 import { GasPlanet } from "@/frontend/universe/planets/gasPlanet/gasPlanet";
 
-import { ItemPool } from "@/utils/itemPool";
+import { ItemPool } from "@/utils/dataStructures/itemPool";
 
 import { enablePhysics } from "../utils";
